Fix broken test
