+++ src/utils/useDidMountEffect.ts
@@ -1,12 +1,13 @@
 import { useEffect, useState } from 'react';
 
 const useDidMountEffect = (func: () => void, deps: Array<unknown>): void => {
-  const [didMount, setDidmount] = useState(false);
+  const [didMount, setDidMount] = useState(false);
   useEffect(() => {
+    console.log('effect');
     if (didMount) {
       func();
     } else {
-      setDidmount(true);
+      setDidMount(true);
     }
   }, deps);
 };
